flask_segmentation/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
efficientnet-pytorch==0.7.1
filelock==3.13.3
Flask==3.0.2
fonttools==4.50.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pretrainedmodels==0.7.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
segmentation-models-pytorch==0.3.3
six==1.16.0
sympy==1.12
timm==0.9.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
